代码优化

+++ updateHosts.py
@@ -44,9 +44,9 @@ def get_config():
 
 
 def backup_hosts():
-    if not os.path.isfile(hosts_folder + 'backup_hosts_original_by_updateHosts'):
-        if os.path.isfile(hosts_folder + 'hosts'):
-            shutil.copy(hosts_folder+'hosts', hosts_folder+'backup_hosts_original_by_updateHosts')
+    if (not os.path.isfile(hosts_folder + 'backup_hosts_original_by_updateHosts')) and \
+            os.path.isfile(hosts_folder + 'hosts'):
+        shutil.copy(hosts_folder+'hosts', hosts_folder+'backup_hosts_original_by_updateHosts')
     if os.path.isfile(hosts_folder + 'hosts'):
         shutil.copy(hosts_folder+'hosts', hosts_folder+'backup_hosts_last_by_updateHosts')
